Created getEntryURL and getEntryTagURL in domain
