+++ src/domain/section.ts
@@ -66,3 +66,54 @@ export const sectionsConfig: Record<SectionType, SectionConfig> = {
     showFeaturedImage: true
   }
 }
+
+/**
+ * Obtiene la ruta (slug) para una sección en un idioma concreto
+ */
+export function getSectionRoute(section: SectionType, locale: UILanguages): string {
+  return sectionsConfig[section].routes[locale]
+}
+
+/**
+ * Get the url of a section in a specific locale
+ * @param section section to get the url for
+ * @param locale locale to get the url for
+ * @returns route string for the section in the locale
+ */
+export function getURLForSection(
+  section: SectionType,
+  locale: UILanguages
+): string {
+  const route = getSectionRoute(section, locale)
+  return `/${locale}/${route}`
+}
+
+/**
+ * Gets the url of a specific entry in a section and locale
+ * @param section section of the content
+ * @param locale locale of the entry
+ * @param slug clean id of the entry
+ * @returns url string for the entry
+ */
+export function getEntryURL(
+  section: SectionType,
+  locale: UILanguages,
+  slug: string
+): string {
+  return `${getURLForSection(section, locale)}/${slug}`
+}
+
+/**
+ * Gets the url of a specific tag in a section and locale
+ * @param section section of the content
+ * @param locale locale of the entry
+ * @param tag tag to get the url for
+ * @returns url string for the tag in the section and locale
+ */
+export function getEntryTagURL(
+  section: SectionType,
+  locale: UILanguages,
+  tag: string
+): string {
+  return `${getURLForSection(section, locale)}/tags/${tag}`
+}
