+++ tests/support/ui/shared/NavigationPaths.ts
@@ -1,6 +1,6 @@
 import type { UILanguages } from '@i18n/ui'
-import type { SectionType } from '@domain/section'
-import { getURLForSection } from '@utils/sections'
+import { type SectionType, getEntryTagURL, getEntryURL, getURLForSection } from '@domain/section'
+import {  } from '@utils/sections'
 import type { Page } from '@playwright/test'
 import { step } from '@tests/fixtures'
 import { mockThirdParty } from '@tests/e2e/helpers/mockThirdParty'
@@ -18,11 +18,11 @@ export function contentListPath(section: SectionType, locale: UILanguages) {
 }
 
 export function contentDetailsPath(section: SectionType, locale: UILanguages, slug: string) {
-  return `${contentListPath(section, locale)}/${slug}`
+  return getEntryURL(section, locale, slug)
 }
 
 export function contentTagsPath(section: SectionType, locale: UILanguages, tag: string) {
-  return `${contentListPath(section, locale)}/tags/${tag}`
+  return getEntryTagURL(section, locale, tag)
 }
 
 export const visit = <T>(
