+++ tests/unit/domain/section.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { getSectionRoute, getURLForSection } from '@utils/sections'
+import { getSectionRoute, getURLForSection, getEntryURL, getEntryTagURL } from '@domain/section'
 
 describe('section route helpers', () => {
   it('getSectionRoute returns correct route for locale', () => {
@@ -12,5 +12,15 @@ describe('section route helpers', () => {
     expect(getURLForSection('blog', 'es')).toBe('/es/blog')
     expect(getURLForSection('talk', 'es')).toBe('/es/charla')
     expect(getURLForSection('work', 'en')).toBe('/en/work')
+  })
+
+  it('getEntryURL builds full url for entry with locale prefix', () => {
+    expect(getEntryURL('blog', 'es', 'my-post')).toBe('/es/blog/my-post')
+    expect(getEntryURL('talk', 'en', 'my-talk')).toBe('/en/talk/my-talk')
+  })
+
+  it('getEntryTagURL builds full url for tag with locale prefix', () => {
+    expect(getEntryTagURL('blog', 'es', 'my-tag')).toBe('/es/blog/tags/my-tag')
+    expect(getEntryTagURL('talk', 'en', 'my-tag')).toBe('/en/talk/tags/my-tag')
   })
 })
